Added reset methods to Moonlake_Model_Condition.
This changes makes it possible to reuse an allready
used condition object. This results in lesser count
of needed objects. There is a reset function for
each type of condition, so you can reset only one
type of condition. But keep in mind, that this can
cause hard to find bugs. Better you use resetAll()
to reset the hole condition object. After that it
is in the same state a like after instanciating it.