Fixed bug in Moonlake_Model_MySQLConnector.
This bug can cause multiple usage of the same id
for different queries or prepared statements. This 
can occure if there are more than one query or
prepared statement and an older one is deleted. 
For instance if you have two queries, the first
one will have the id 1 and the second one the id 2
Then if you free the first one (with id 1), the
next query will also have the id 2.